authentication_service/requirements.txt
aerich==0.6.3
aiosqlite==0.17.0
alembic==1.8.0
anyio==3.6.1
asgiref==3.5.2
attrs==21.4.0
bcrypt==3.2.2
certifi==2022.5.18.1
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.3
databases==0.6.0
dictdiffer==0.9.0
distlib==0.3.4
fastapi==0.78.0
filelock==3.7.1
greenlet==1.1.2
h11==0.13.0
identify==2.5.1
idna==3.3
iniconfig==1.1.1
iso8601==1.0.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
nodeenv==1.6.0
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.19.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
pyparsing==3.0.9
pypika-tortoise==0.1.5
pytest==7.1.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.37
starlette==0.19.1
tld==0.12.6
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
tortoise-orm==0.19.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
virtualenv==20.14.1
